Allow compilation with aeson 2.2.0.0.
This version of aeson removes Data.Aeson.Internal, but exports the
things yaml needed from it in Data.Aeson.Types.

CPP has been used to make it work with both 2.2 and earlier versions.
However, I was only able to test with aeson 2.1; attempts to constrain
the build to earlier versions did not produce a viable build path.

Closes #219.